added dynamic allocation of the tokens array to support larger input files. bumped the read_pos var for the parser to u16 from u8 to prevent int overflow issues. passing large.ini parsing test (which ensures the final item is captured)

+++ src/parser.zig
@@ -3,7 +3,7 @@ const token_types = @import("token.zig");
 const iniConfig = @import("ini-config.zig");
 
 pub const Parser = struct{
-    read_pos: u8,
+    read_pos: u16,
 
     tokens: *[]token_types.Token,
     parseMap: iniConfig.IniConfig,
